Remove extra hash
c3 already hashes even ulong, it doesn't just use the ulong itself to
calculate the index. Not sure if this is true for the rust version,
which is why it tries a "prehash" strategy. But c3 is faster even doing
the "proper" hashing strategy.

I believe zig also ends up hashing u64, so its prehash should also be
removed.

After this change, perf is back to even with odin (but not better, as it
was before with the janky set impl). This is pretty impressive, as it's
a complete hashmap implementation, no cut corners. So I guess the janky
set impl is not actually faster, since removing this change accounts for
the performance diff.

One reason the new c3 hashmap doesn't lose perf, even though it has
entries in a separate array, is that the array being searched initially
is basically char[], which is faster than the usz[] or ulong[] being
used before for entries.